Use uint64 for slice (#6)

+++ mass.go
@@ -1,7 +1,7 @@
 package mass
 
 // New returns new pool.
-func New[T any](capacity int) *Mass[T] {
+func New[T any](capacity uint64) *Mass[T] {
 	return &Mass[T]{
 		pool:     make([]T, capacity),
 		capacity: capacity,
@@ -11,7 +11,7 @@ func New[T any](capacity int) *Mass[T] {
 // Mass maintains a pool of objects to avoid excessive number of heap allocations.
 type Mass[T any] struct {
 	pool     []T
-	capacity int
+	capacity uint64
 }
 
 // New returns new object from the pool.
@@ -25,8 +25,8 @@ func (m *Mass[T]) New() *T {
 }
 
 // NewSlice returns new slice of objects from the pool.
-func (m *Mass[T]) NewSlice(n int) []T {
-	if n > len(m.pool) {
+func (m *Mass[T]) NewSlice(n uint64) []T {
+	if n > uint64(len(m.pool)) {
 		if n2 := n << 1; n2 > m.capacity {
 			m.capacity = n2
 		}
